prep tests for i2

+++ sysbench.sh
@@ -1,5 +1,5 @@
 #! /bin/bash
-export DEBIAN_FRONTEND=noninteractive 
+export DEBIAN_FRONTEND=noninteractive
 apt-get update; apt-get install sysbench mysql-server -y -q
 
 
@@ -14,11 +14,6 @@ MAX_PRIMES=200000
 TEST_TIMEOUT=240
 OLTP_ROWS_NUM=1000000
 FILE_SIZE=40G
-
-MAX_PRIMES=2000
-TEST_TIMEOUT=3
-OLTP_ROWS_NUM=1000
-FILE_SIZE=40M
 
 
 CPU_TESTS=$(echo "1;$CPU_COUNT/2;$CPU_COUNT" | bc  | grep -v  '^0$' | sort | uniq)
@@ -44,7 +39,7 @@ function cpu_prime {
 function oltp {
     THREAD_COUNT=$1
     RESULT_FILE="$RESULT_DIR/${FUNCNAME[0]}.${THREAD_COUNT}.txt"
-   
+
     COMMAND="sysbench --test=oltp --oltp-table-size=$OLTP_ROWS_NUM --mysql-db=test --mysql-user=root --num-threads=$THREAD_COUNT run"
     echo $COMMAND
     echo "# $COMMAND" > $RESULT_FILE
@@ -114,44 +109,50 @@ function io_seqrd {
 dest=$HOME/sb_files
     mkdir -p $dest
     pushd $dest
-    sysbench --test=fileio --file-total-size=$FILE_SIZE prepare
+    if [[ -z "$NO_DISK_TEST" ]] then
+        sysbench --test=fileio --file-total-size=$FILE_SIZE prepare
+    fi
 
-    for tc in $CPU_TESTS 
+    for tc in $CPU_TESTS
     do
-    	cpu_prime $tc
-    	oltp $tc
-    	for bs in $BS_TESTS
-    	do
-    		io_rndrd $tc $bs root
-    		io_rndwr $tc $bs root 
-    		io_seqrd $tc $bs root 
-    		io_seqwr $tc $bs root
-    	done
+        cpu_prime $tc
+        oltp $tc
+        if [[ -z "$NO_DISK_TEST" ]] then
+            for bs in $BS_TESTS
+            do
+                io_rndrd $tc $bs root
+                io_rndwr $tc $bs root
+                io_seqrd $tc $bs root
+                io_seqwr $tc $bs root
+            done
+        fi
     done
 
     popd
     rm -rf $dest
 
-if (mount | grep "/dev/vdb on /mnt")
-then
-    dest=/mnt/sb_files
-        mkdir -p $dest
-        pushd $dest
-        sysbench --test=fileio --file-total-size=$FILE_SIZE prepare
+if [[ -z "$NO_DISK_TEST" ]] then
+    if (mount | grep "/dev/vdb on /mnt")
+    then
+        dest=/mnt/sb_files
+            mkdir -p $dest
+            pushd $dest
+            sysbench --test=fileio --file-total-size=$FILE_SIZE prepare
     
-        for tc in $CPU_TESTS
-        do
-        	for bs in $BS_TESTS
-        	do
-        		io_rndrd $tc $bs ephemeral
-        		io_rndwr $tc $bs ephemeral 
-        		io_seqrd $tc $bs ephemeral 
-        		io_seqwr $tc $bs ephemeral
-        	done
-        done
+            for tc in $CPU_TESTS
+            do
+                for bs in $BS_TESTS
+                do
+                    io_rndrd $tc $bs ephemeral
+                    io_rndwr $tc $bs ephemeral
+                    io_seqrd $tc $bs ephemeral
+                    io_seqwr $tc $bs ephemeral
+                done
+            done
     
-        popd
-        rm -rf $dest
+            popd
+            rm -rf $dest
+    fi
 fi
 
 python sysbench_parse.py $RESULT_DIR > sb_result.json
